added restrictions to upate and delete published quiz also Added validator for various checks

+++ Project/QuizManager/src/controllers/quiz.ts
@@ -1,6 +1,7 @@
 import { Request, Response, NextFunction } from "express"
 import Quiz from "../models/quiz";
 import ProjectError from "../helper/error";
+import { validationResult } from "express-validator/src/validation-result";
 
 interface ReturnResponse{
     status:"success" | "error",
@@ -10,6 +11,15 @@ interface ReturnResponse{
 
 const createQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     try {
+
+        const validationError=validationResult(req);
+        if(!validationError.isEmpty()){
+            const err=new ProjectError("Validation Failed");
+            err.statusCode=422; 
+            err.data=validationError.array();
+            throw err;
+        }
+
         const created_by=req.userId;
         const name=req.body.name;
         const questions_list=req.body.questions_list;
@@ -51,7 +61,16 @@ const getQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     }
 }
 
-const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{    try {
+const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{    
+    try {
+        const validationError=validationResult(req);
+        if(!validationError.isEmpty()){
+            const err=new ProjectError("Validation Failed");
+            err.statusCode=422; 
+            err.data=validationError.array();
+            throw err;
+        }
+
         const quizId=req.body._id;
         const quiz=await Quiz.findById(quizId);
         if(!quiz){
@@ -64,6 +83,12 @@ const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{    try
             err.statusCode=403;
             throw err;
         }
+        if(quiz.is_published){
+            const err=new ProjectError("You cannot update published quiz");
+            err.statusCode=405;
+            throw err;
+        }
+
         quiz.name=req.body.name;
         quiz.questions_list=req.body.questions_list;
         quiz.answers=req.body.answers;
@@ -90,7 +115,11 @@ const deleteQuiz=async (req:Request, res:Response, next:NextFunction) =>{
             err.statusCode=403;
             throw err;
         }
-
+        if(quiz.is_published){
+            const err=new ProjectError("You cannot delete published quiz");
+            err.statusCode=405;
+            throw err;
+        }
         await Quiz.deleteOne({_id:quizId});
         const resp:ReturnResponse={status:"success", message:"Quiz deleted successfully", data:{}};
         res.status(200).send(resp);
